Update HardeningDebianv1.sh
need an alternative to wheel group for Debian I think it may be Arch only by default, not quite sure what solution would be good in context of that part of the script?  Perhaps creaste a wheel group? Would have to research exactly what that group is and how it functions ( under which rules )